Use Staking/ClaimedRewards as updated API for claims already performed
The current version of Substrate relies on a new API (`Staking/ClaimedRewards`) to check the claims that have been performed by validators in any era.

This commit introduces the support for this API that is used in conjunction with the previous one (`Staking/ErasValidatorRewards`) to build a map of claims.

The previous API (now deprecated) is not removed so that the tool can still support older chains, but the `claimed_rewards` field in the output has been renamed as `legacy_claimed_rewards` to reflect the changes in the Substrate framework.

In the long term, the old API could probably be safely removed.