Support chains without batch extrinsic
The previous version of the code only supported chains with the `Batch` extrinsic available to perform all the payouts with a single call.

This commit adds support also for chains without the `Batch` feature (`Utility` pallet).

Batching is used only if available, single extrinsics are created otherwise, one for each payout (validator|era).